chore(popup): Added playwright tests
✅ Closes: COMUI-4071

chore(popup): Refactored e2e test file

✅ Closes: COMUI-4071
