+++ packages/genesys-spark-components/src/components/stable/gux-popup/tests/gux-popup.e2e.playwright.ts
@@ -1,0 +1,10 @@
+import { checkRenders, test } from '@test/playwrightTestUtils';
+import { renderConfigs } from './gux-popup.common';
+
+test.describe('gux-popup', () => {
+  test.describe('#render', () => {
+    checkRenders({
+      renderConfigs
+    });
+  });
+});
